Address WASM registration review feedback (#37)
Move WASM preparation details into a sibling registry module so
`loader.rs` keeps only the registration orchestration and persistence
boundary.

Add a regression test covering credential publication for successful
WASM registration and protecting the shared credential registry when
registration is rejected.
