+++ src/tools/registry/wasm_registration_tests.rs
@@ -1,16 +1,20 @@
 //! Regression tests for proactive WASM schema publication during registration.
 
+use std::sync::Arc;
+
 use chrono::Utc;
 use rstest::rstest;
 use uuid::Uuid;
 
-use super::{ToolRegistry, WasmFromStorageRegistration};
+use super::{ToolRegistry, WasmFromStorageRegistration, WasmToolRegistration};
 use crate::llm::ToolDefinition;
+use crate::secrets::CredentialMapping;
+use crate::testing::credentials::test_secrets_store;
 use crate::testing::{github_wasm_artifact, metadata_test_runtime};
 use crate::tools::wasm::storage::{NativeWasmToolStore, ToolKey};
 use crate::tools::wasm::{
-    StoredCapabilities, StoredWasmTool, StoredWasmToolWithBinary, ToolStatus, TrustLevel,
-    WasmStorageError,
+    Capabilities, HttpCapability, SharedCredentialRegistry, StoredCapabilities, StoredWasmTool,
+    StoredWasmToolWithBinary, ToolStatus, TrustLevel, WasmError, WasmStorageError,
 };
 
 #[rstest]
@@ -54,9 +58,91 @@ async fn register_wasm_from_storage_recovers_guest_schema_for_absent_or_placehol
     assert_real_github_schema(definition);
 }
 
+#[tokio::test]
+async fn register_wasm_persists_credentials_only_after_successful_registration() {
+    let credential_registry = Arc::new(SharedCredentialRegistry::new());
+    let registry = ToolRegistry::new().with_credentials(
+        Arc::clone(&credential_registry),
+        Arc::new(test_secrets_store()),
+    );
+    let runtime = metadata_test_runtime().expect("create metadata test runtime");
+    let wasm_binary = github_wasm_bytes();
+
+    let rejected = registry
+        .register_wasm(registration_with_credential(
+            "http",
+            &wasm_binary,
+            &runtime,
+            "rejected_token",
+            "rejected.example.com",
+        ))
+        .await;
+    assert!(
+        matches!(rejected, Err(WasmError::ConfigError(message)) if message == "tool registration rejected"),
+        "protected tool name should reject registration"
+    );
+    assert!(
+        credential_registry
+            .find_for_host("rejected.example.com")
+            .is_empty(),
+        "rejected registrations must not publish credential mappings"
+    );
+
+    registry
+        .register_wasm(registration_with_credential(
+            "github",
+            &wasm_binary,
+            &runtime,
+            "accepted_token",
+            "api.example.com",
+        ))
+        .await
+        .expect("successful registration should persist credentials");
+
+    let mappings = credential_registry.find_for_host("api.example.com");
+    assert_eq!(mappings.len(), 1);
+    assert_eq!(mappings[0].secret_name, "accepted_token");
+    assert!(
+        credential_registry
+            .find_for_host("rejected.example.com")
+            .is_empty(),
+        "failed registrations should leave the credential registry unchanged"
+    );
+}
+
 fn github_wasm_bytes() -> Vec<u8> {
     let wasm_path = github_wasm_artifact().expect("build or find github WASM artifact");
     std::fs::read(wasm_path).expect("read github wasm artifact")
+}
+
+fn registration_with_credential<'a>(
+    name: &'a str,
+    wasm_bytes: &'a [u8],
+    runtime: &'a Arc<crate::tools::wasm::WasmToolRuntime>,
+    secret_name: &str,
+    host_pattern: &str,
+) -> WasmToolRegistration<'a> {
+    WasmToolRegistration {
+        name,
+        wasm_bytes,
+        runtime,
+        capabilities: capabilities_with_credential(secret_name, host_pattern),
+        limits: None,
+        description: Some("Credential persistence test tool"),
+        schema: Some(serde_json::json!({
+            "type": "object",
+            "properties": {}
+        })),
+        secrets_store: None,
+        oauth_refresh: None,
+    }
+}
+
+fn capabilities_with_credential(secret_name: &str, host_pattern: &str) -> Capabilities {
+    Capabilities::default().with_http(HttpCapability::default().with_credential(
+        secret_name,
+        CredentialMapping::bearer(secret_name, host_pattern),
+    ))
 }
 
 fn assert_real_github_schema(definition: ToolDefinition) {
